lab6: 4.2.3 done

+++ src/lab6/fs/virtio.c
@@ -27,12 +27,12 @@ void virtio_blk_feature_init() {
 
 void virtio_blk_queue_init() {
     virtio_blk_ring.num = VIRTIO_QUEUE_SIZE;
-    
+
     uint64_t size_of_descs = VIRTIO_QUEUE_SIZE * sizeof(struct virtio_desc);
     uint64_t size_of_avail = sizeof(struct virtio_avail);
     uint64_t size_of_used = sizeof(struct virtio_used);
 
-    uint64_t pages = alloc_pages(3);
+    uint64_t pages = (uint64_t)alloc_pages(3);
     virtio_blk_ring.desc = (struct virtio_desc*)(pages);
     virtio_blk_ring.avail = (struct virtio_avail*)(pages + PGSIZE);
     virtio_blk_ring.used = (struct virtio_used*)(pages + 2*PGSIZE);
@@ -76,7 +76,7 @@ void virtio_blk_cmd(uint32_t type, uint32_t sector, void* buf) {
     virtio_blk_ring.desc[1].addr = virt_to_phys((uint64_t)buf);
     virtio_blk_ring.desc[1].len = VIRTIO_BLK_SECTOR_SIZE;
     if (type == VIRTIO_BLK_T_IN) {
-        virtio_blk_ring.desc[1].flags = VIRTQ_DESC_F_WRITE | VIRTQ_DESC_F_NEXT; 
+        virtio_blk_ring.desc[1].flags = VIRTQ_DESC_F_WRITE | VIRTQ_DESC_F_NEXT;
     } else {
         virtio_blk_ring.desc[1].flags = VIRTQ_DESC_F_NEXT;
     }
